couple fixes and removals of console log
